+++ wave/src/com/game/main/FastEnemy.java
@@ -14,7 +14,7 @@ public class FastEnemy extends GameObject {
 	}
 	
 	public Rectangle getBounds() {
-		return new Rectangle(x, y, 16, 16);
+		return new Rectangle((int) x, (int) y, 16, 16);
 	}
 
 	public void tick() {
@@ -31,13 +31,13 @@ public class FastEnemy extends GameObject {
 		 * Fire off Trail objects that will slowly fade out in their place as
 		 * the BasicEnemy moves around the screen.
 		 */
-		new Trail(x, y, ID.Trail, Color.cyan, 16, 16, 0.01f, handler);
+		new Trail((int) x, (int) y, ID.Trail, Color.cyan, 16, 16, 0.01f, handler);
 		
 	}
 
 	public void render(Graphics g) {
 		g.setColor(Color.cyan);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int) x, (int) y, 16, 16);
 	}
 
 }
