+++ wave/src/com/game/main/HUD.java
@@ -10,7 +10,7 @@ public class HUD {
 	 * it available as HUD.HEALTH without having to initialize a HUD
 	 * object.
 	 */
-	public static int HEALTH = 100;
+	public static float HEALTH = 100;
 	
 	private int score = 0;
 	private int level = 1;
@@ -30,7 +30,7 @@ public class HUD {
 		 */
 		g.setColor(Color.getHSBColor((1f * HEALTH) / 360, 1f, 1f));
 		
-		g.fillRect(16, 16, HEALTH * 2, 32);
+		g.fillRect(16, 16, (int) HEALTH * 2, 32);
 		g.setColor(Color.white);
 		g.drawRect(16, 16, 200, 32);
